cloud_functions/requirements.txt
absl-py
google-api-python-client
google-auth
google-ads
google-cloud-functions
google-cloud-secret-manager
pandas
parameterized
paramiko
